Linux wifi adapter setup/requirements.txt
RPi.GPIO
st7789
